+++ mods/BrewLAN/modules/BrewUI/spawnmenu/mod_info.lua
@@ -1,8 +1,8 @@
 name = "BrewLAN UI: Spawn Menu"
-uid = "25D57D85-7D84-27HT-A501-BR3WUISM00002"
-version = 0.2
-copyright = "© 2020–2022 Sean Wheeldon"
-description = "Gives the ALT-F2 cheatspawn menu mod filters. Functions on any version of Supreme Commander or Forged Alliance."
+uid = "25D57D85-7D84-27HT-A501-BR3WUISM00008"
+version = 0.8
+copyright = "© 2020–2023 Sean Wheeldon"
+description = "Overhauls the cheat spawn menu."
 author = "Balthazar"
 icon = "/mods/BrewLAN/modules/BrewUI/spawnmenu/BrewUI.png"
 selectable = true